Roll netfx SqlConfigurableRetryLogicManager.LoadType.cs into SqlConfigurableRetryLogicLoader
